pyproject.toml
[tool.poetry]
name = "babybottlelog"
version = "0.1.0"
description = ""
authors = ["Maarten", "Ling"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.41.1"
jupyter = "^1.1.1"
plotly = "^5.24.1"
black = "^24.10.0"
ruff = "^0.8.4"
isort = "^5.13.2"
azure-storage-blob = "^12.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
